requirements.txt
requests
python-telegram-bot
python-dotenv
pandas
schedule
groq
